+++ magick/cache_io.h
@@ -18,8 +18,8 @@ extern "C" {
 extern Export ClassType
   GetCacheClassType(CacheHandle);
 
-extern Export size_t
-  GetCacheMemory(size_t);
+extern Export off_t
+  GetCacheMemory(off_t);
 
 extern Export unsigned int
   GetCacheThreshold(),
